fix: widen telegram_id to BIGINT and stop DB errors masquerading as swap errors (#282)
Comprehensive companion to the narrow telegram_id fix. Addresses both
the root cause (new Telegram user IDs overflow the int4 column) and the
amplifier (every non-swap exception gets rendered to users as a swap
error because the global error handler pattern-matches the stringified
SQL).

Root cause — telegram_id INT4 overflow:

- bot/models/user.py: User.telegram_id Integer -> BigInteger.
- database/db.py: add idempotent _widen_user_telegram_id() that runs
  ALTER TABLE users ALTER COLUMN telegram_id TYPE BIGINT on Postgres
  when the column is still 'integer'. SQLite stores INTEGER as 64-bit
  so no migration is needed there.
- api-ts/src/db/schema/users.ts: telegramId switched from integer() to
  bigint({ mode: 'number' }) so the Drizzle schema stays aligned with
  Postgres. 'number' mode is safe — Telegram IDs fit well within
  Number.MAX_SAFE_INTEGER (2^53 - 1).

Error-handler amplifier:

bot/utils/errors.py previously ran detect_error_code() against the
stringified exception, including SQLAlchemy's embedded SQL text. The
failing INSERT contains the column name 'default_slippage', and the
regex /slippage/i matched that substring, so every DB failure was
reported to users as "Price moved too much during swap. Increase
slippage tolerance or try again."

detect_error_code() now short-circuits to UNKNOWN_ERROR when the
message contains DB/ORM markers ([sql:, psycopg2., sqlalchemy.exc,
IntegrityError, etc.). detect_error_code_from_exception() additionally
checks the exception's module (psycopg2, sqlalchemy, asyncpg, aiosqlite)
before falling back to string matching — so DB exceptions are filtered
out at the class level, not just by their message text.

Legit swap-slippage errors (e.g. Exception("slippage exceeded")) still
resolve to SLIPPAGE_TOO_HIGH. Verified with a smoke test covering all
four cases (DB error with column name; real slippage message; psycopg2
class; generic exception with slippage keyword).

This is the comprehensive variant. A narrow companion PR touches only
bot/models/user.py and database/db.py for teams that want minimum
review surface. Either merge alone resolves the /start blockage.

+++ bot/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, Boolean, ForeignKey, Text, Float
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, Boolean, ForeignKey, Text, Float
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from database.db import Base
@@ -7,9 +7,10 @@ from database.db import Base
 class User(Base):
     """Telegram user model."""
     __tablename__ = "users"
-    
+
     id = Column(Integer, primary_key=True)
-    telegram_id = Column(Integer, unique=True, nullable=True, index=True)
+    # Telegram user IDs routinely exceed 2^31 (INT4 max) as of 2024-2026; must be BIGINT.
+    telegram_id = Column(BigInteger, unique=True, nullable=True, index=True)
     whatsapp_id = Column(String(255), unique=True, nullable=True, index=True)
     discord_id = Column(String(100), unique=True, nullable=True, index=True)
     discord_username = Column(String(255), nullable=True)
